backend/classification/requirements.txt
mlflow==2.19.0
dvc==3.58.0
dvc-azure==3.1.0
scikit-learn==1.5.2
pandas==2.2.3
numpy>=1.26.0,<2.1
seaborn==0.13.2
matplotlib==3.10.0
pytest==8.3.4
PyYAML==6.0.2 
python-dotenv==1.0.1
nannyml==0.13.0 
scipy==1.15.1
nbformat>=4.2.0

azure-storage-blob=<12.24.1
